refactor: Moving from stores to page context.
