+++ src/admin/pages/User/Context/type.ts
@@ -1,0 +1,6 @@
+import { User } from '@shared/types';
+import { SWRResponse } from 'swr';
+
+export type UserContext = {
+  getUsers: () => SWRResponse<User[]>;
+};
